requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
filelock==3.9.0
identify==2.5.18
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
marshmallow==3.19.0
marshmallow-dataclass==8.5.11
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.0
platformdirs==2.6.2
pre-commit==3.1.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing-inspect==0.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
